Update patch set 1
Patch Set 1:

I'm not sure if this actually pulls in the subrepo. This is what you did by making it part of the workspace? It doesn't pull it in for me.

Is it because you're not actually depending on anything from it yet (cpplint.py)?

T think that's whats happening.

Patch-set: 1